First pass for translating the code from Go and JS projects
We are using the following two code-bases as reference:
- https://github.com/open-spaced-repetition/fsrs4anki/blob/50aed4be1ab215b226bcef91447961df120339d5/fsrs4anki_scheduler.js
-
https://github.com/open-spaced-repetition/go-fsrs/blob/70232f222d0c0ef7523ffa7ceb3de91feddf6030/fsrs.go

The current code is a translation of the two code-bases above into
Clojure. Whenever there is a conflict in the two code-bases, we give a
preference to the JS version.

All the formulae are double checked with FSRS v4 from: https://github.com/open-spaced-repetition/fsrs4anki/wiki/Free-Spaced-Repetition-Scheduler

Next steps:
1. Add tests to ensure that the code is working correctly
2. Refactor the code to be cleaner and simpler to reason about
3. Refactor the code, if possible, to be platform independent. (In
   this case, we might end up renaming the entire project to
   cljc-fsrs. But this will only be done if time permits)