Fix: Rely on ->rating for the numeric value of each rating
We do not store the weight of the rating in the weights param, we
store it in the ->rating map and should take it from there.

Note that when initializing stability, we are taking the stability
parameter from the `weights` array, based on the rating. This is not a
mistake.